ai-onboarding-agent/requirements.txt
PyGithub==1.59.1
kubernetes==28.1.0
jinja2==3.1.2
openai==1.3.7
python-dotenv==1.0.0